+++ src/ProjectTypeResolver.php
@@ -19,7 +19,10 @@ class ProjectTypeResolver
     /**
      * The key from the composer configuration which contains other configuration.
      */
-    public const COMPOSER_CONFIG_KEY = 'youwe-testing-suite';
+    public const COMPOSER_CONFIG_KEYS = [
+        'youwe-testing-suite',
+        'mediact-testing-suite'
+    ];
 
     /**
      * The key in the configuration, which determines the overwrite for the type.
@@ -38,6 +41,8 @@ class ProjectTypeResolver
         'alumio-project'  => 'alumio',
         'laravel-project' => 'laravel',
     ];
+
+    public const DEFAULT_PROJECT_TYPE = 'default';
 
     /**
      * Constructor.
@@ -60,10 +65,12 @@ class ProjectTypeResolver
     {
         $config = $this->composer->getConfig();
 
-        if ($config->has(static::COMPOSER_CONFIG_KEY)) {
-            $configNode = $config->get(static::COMPOSER_CONFIG_KEY);
-            if (isset($configNode[static::COMPOSER_CONFIG_TYPE_KEY])) {
-                return $configNode[static::COMPOSER_CONFIG_TYPE_KEY];
+        foreach (static::COMPOSER_CONFIG_KEYS as $key) {
+            if ($config->has($key)) {
+                $configNode = $config->get($key);
+                if (isset($configNode[static::COMPOSER_CONFIG_TYPE_KEY])) {
+                    return $configNode[static::COMPOSER_CONFIG_TYPE_KEY];
+                }
             }
         }
 
@@ -71,6 +78,6 @@ class ProjectTypeResolver
 
         return array_key_exists($packageType, $this->mapping)
             ? $this->mapping[$packageType]
-            : 'default';
+            : self::DEFAULT_PROJECT_TYPE;
     }
 }